refactor : update Partner Ui to depend on the interface insted of the implementation .
